backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
celery==5.3.4
Pillow==10.1.0
python-magic==0.4.27